changed file structure to include app package, as well as implemented firebase writer method using springdata
