+++ src/main/java/ai/classlynk/app/AppConfig.java
