+++ src/main/java/ai/classlynk/entity/ClassBundle.java
@@ -1,4 +1,4 @@
-package entity;
+package ai.classlynk.entity;
 
 import java.util.List;
