+++ src/main/java/ai/classlynk/entity/Timetable.java
@@ -1,4 +1,4 @@
-package entity;
+package ai.classlynk.entity;
 
 import java.util.List;
 import java.util.Map;
