+++ src/main/java/ai/classlynk/use_case/generate_timetable/TimetableGeneratorDataAccessInterface.java
@@ -1,4 +1,4 @@
-package use_case.generate_timetable;
+package ai.classlynk.use_case.generate_timetable;
 
 
 public interface TimetableGeneratorDataAccessInterface {
